+++ api/smart_template/course_content_info.py
