+++ api/smart_template/live_template_delete.py
