+++ api/smart_template/live_template_detail.py
