+++ test_case/smart_template_case/test_smart_live_template_bank_post.py
@@ -8,9 +8,9 @@ info:测试新建智能直播模板组，然后连接数据库删除它
 
 import pytest
 
-from Common.EnvConfig import host, headers
-from Common.MysqlConfig import db_study_center
-from api.SmartTemplate.smart_live_template_bank_post import smart_live_template_bank_post
+from common.env_config import host, headers
+from common.mysql_config import db_study_center
+from api.smart_template.smart_live_template_bank_post import smart_live_template_bank_post
 
 
 def test_1(db):
